// Frontend/src/app/models/fmeasures.ts
export interface FlootMeasurements {
    id_medicion?: number;
    masa: number;
    estatura: number;
    altura_ojo: number;
    altura_hombro: number;
    altura_codo: number;
    altura_entre_pierna: number;
    profundidad_cuerpo_de_pie: number;
    anchura_cadera_pie: number;
    id_trabajador_pertenece: number;
  }
  